bios: mark .note.gnu.build-id as non-loadable but don't disable it
If the linker is told to inject a .note.gnu.build-id section using
--build-id, it will be included in bios.elf but will be marked
non-loadable similar to the other kinds of metadata that already exists.

Marking it non-loadable ensures that it won't be placed anywhere in the
address space, so it will be excluded from the bios.bin image and won't
cause any issues due to the section overlapping with .text:

    ld: section .note.gnu.build-id LMA [0000000000000000,0000000000000023] overlaps section .text LMA [0000000000000000,000000000000519b]

or in later binutils versions:

    ld: error: no memory region specified for loadable section `.note.gnu.build-id'

Relates to #1020, #1042.