XilinxVivadoToolchain: support copying dependencies to the build dir
This introduces a flag called `copy_deps` which, when specified,
attempts to properly copy all required dependencies for the Xilinx
Vivado build into the target build directory. This is particularly
useful for shipping the generated Verilog and not have it reference
system dependencies, or to integrate with scripts for building the
bitstreams on a remote machine.

This mechanism is by no means perfect, there might still be absolute
paths pointed to by included Tcl scripts, etc. However it's a good
start and the proper way to handle dependencies such as the
pythondata-cpu-* files. With these changes, developers can simply
adjust their own scripts to also only have relative paths and it'll
likely just work.

Signed-off-by: Leon Schuermann <leon@is.currently.online>