README/litex_setup: Remove reference to LiteVideo to encourage use of LiteX's VideoTerminal/Out core.
LiteVideo is not longer maintained, does not have CI and is messy (code is ~10 years old where we were
still experimenting the innovative approach with Migen). The core is kept since can be useful as reference
for Video Input and for projects using it but it is not recommended for new designs.