litex_sim: add ZeroMQ/JSON-based simctrl interface module
Adds a module implementing the "simctrl"-style control & status
interface for the LiteX simulation. It uses a ZeroMQ/JSON-based
transport for the control, status and module messages. However,
modules are free to define their own format.

The passed messages make use of ZeroMQ multipart messages, for
instance to carry payloads to simctrl-supported modules in the
simulation. Thus, this module takes messages which contain a "simctrl
header" and a payload forwarded to the destination module.

Furthermore, the module implements some basic standalone status &
control functionality. It can be used to query the current wall clock
time, simulation time and instantiated module sessions. It can
furthermore halt and resume the simulation.

Signed-off-by: Leon Schuermann <leon@is.currently.online>