litex_sim: add inter-module message passing mechanism
This introduces a generic, flexible and efficient inter-module message
passing mechanism to allow for more advanced module behavior and
interactions, as well as outside control of module behavior.

Possible applications for this mechanism are outside control of
emulated peripherals such as GPIO, simulated hardware attached to
multiple independent subsystems (e.g. GPIO + SPI) and integration of
the LiteX simulation in automated systems such as CI.

The message passing mechanism does not impose any restrictions on the
data exchanged between modules and each module is free to define its
own interfaces through opcodes >= 256. Opcodes < 256 are reserved for
globally defined operations.

These global operations include auto-announcement of instantiated
module sessions in the simulation for discovery and a mechanism for
interacting with outside applications (called "simctrl"). These
operations are well documented in `modules.h`.

This further introduces a few functions which expose basic information
about and control of the simulation, such as querying the current
simulation time and state, and requesting the simulation to halt or
resume operation.

Signed-off-by: Leon Schuermann <leon@is.currently.online>