WS2812 off by one error, last LED never updates.
I tested with 8 LEDs and the last LED was never updated. The original code set led_count=Signal(3) and then tested against (nleds-1) for 8 LEDs. This caused a transition to the RST state before processing the last LED. I increased the size of led_count to ensure it was large enough and tested against nleds directly.