litex_json2dts_linux: add config_earlycon constant
This patch checks for a configuration constant named "config_earlycon"
and if present uses it for the "earlycon=" value in the dts
"bootargs".

If the constant is not present, default to the uart defined in the
"csr_bases".

This is useful for the sim.py case, where that configuration wants to
use "earlycon=sbi", while true hardware platforms want the uart from
"csr_bases".

Signed-off-by: Curt Brune <curt@brune.net>