cores/video: new VideoTextGrid component
The VideoTextGrid is similar to the VideoTerminal, but instead of
receiving a stream of characters from an attached UART, it exposes
a csr (YXAC) to allow direct access to the character grid.  Writes
to this register specify the Y coordinate, X coordinate, Attribute,
and Character, encoded in a 32bit value and result in a write to
the character/attribute memory.

The attribute memory currently encodes 3bit RGB foreground and
background colors.  The remaining 2 bits might be useful for
an intensity bit (doubling the colorspace) or some other
feature like h/v flip.

Two additional registers (SCROLLX, SCROLLY) allow the underlying
128x64 character grid to be scrolled within the 80x40 visible
viewport.

Some trickery with the ClockDomainsRenamer was necessary to keep
the register facing side of the text/attr memory.  The "sys" domain
gets renamed to "hdmi" typically, so we put the write side of the
text/attr memory in the "csr" domain and rename that to "sys"
at the same time the "sys" domain is renamed:

vt = ClockDomainsRenamer({ "sys": clock_domain, "csr": "sys" })(vt)