Options to disable verilog generation and to specify doc path
The verilog generation must be disabled if we want to generate doc and
software in a multi-cpu system (not just one multicore cpu, but multiple
(possible different) instances of cpu's in a single device.
- Litex generates the doc and software per instance of a cpu, the user
need to choose proper settings for output directories, of course
- migen platform build will produce the verilog files and builds the
complete set of verilog files.

TODO : check what happens with memories and initialisation of them.