fix false sensitivities in FSM
The re-use of a left hand side (LHS) assignment in a right hand
side (RHS) evaluation inside an `always (@*)` block with non-blocking
operators leads to an infinite loop in some verilog simulators.

In particular, a form such as

fsm.act("A",
   sig.eq(new_value),
   If(sig,
      # ...
   )
)

Causes an infinite loop when `new_value` is not 0. This is because
this gets translated in verilog to

always (@*) begin
   sig <= 1'd0;
   ...
   sig <= new_value;
   if (sig) begin
      ...
   end
end

So what happens is that `sig` gets 0, then, it gets assigned
`new_value`; it is then used by the `if` to decide what to do.

When `new_value` is 0, no change is detected, and there is no
inifnite loop.

However, if `new_value` is 1, the simulator detects the 0->1
transition, and schedules the `always(@*)` for re-evaluation,
because the sensitivity list, which includes `sig` had a transition.

In this case, the simulator (at least some of them) ends up
in an infinite loop, and the simulation is unable to proceed.

The commit attached to this message is just one of several
that fix this motif inside the axi implementations. I have
split this into several successive commits to make it easier
to reject the proposed commits on a case by case basis.