partial fix to AXITimeout implementation
This commit partially fixes AXITimeout. These are the RHS/LHS
loops I could find and fix; however, demonstrably, the system
still oscillates when a timeout is specified. Thus, I am using
the system with `bus_timeout = None` for now, which is fine
because the top-level SoC is just used for validation scaffolding.

What I can't resolve is how to fix the values inside the `response`
list of items that are inserted, somewhat opaquely, into
the "RESPOND" state of the FSM. It looks like the implementation
relies upon being able to "override" the ready/valid values
generated elsewhere through blocking operator semantics, I think.

Anyways, this part of the patch set could be left out since I don't
need it to run my implementation; however, anyone else trying to
configure an SoC using `axi-lite` as a shared bus standard for
the top level and simulating with `xsim` will likely be unable
to simulate past the first write to the shared bus, because it
will trigger an oscillation in the AXITimeout block.