RFC: locate reset vector in custom memory bus
This PR allows the reset vector to be located within a custom memory
bus that is not enumerated by the SoC bus handler.

The use case is when generating a core that is:

1) to be integrated into a third party SoC, where all the memory
bus wiring is handled by a different framework
2) none of the local/automatic busses are desired to be inferred
with the framework.

An example of this usage would be here:

https://github.com/buncram/cram-soc/blob/cram/cram_core.py