POC: Handle un-aligned edge case on AXI upconverter
The AXI upconverter currently always starts the demux aligned
to the width of the destination bus. So, for a 32->64 bit up
converter, it can correctly up-convert a single word-write
with an alignment along 8-byte boundaries. However, "odd"
alignments along 4-byte boundaries are not correctly aligned.

This proof-of-concept patch shows that making the demux starting
address-aware fixes the issue.

However, the way it's done is pretty gross, I'm pretty sure there
is a better way to patch this in.

I'll add more details once the PR is in github.