more fixes
Reveal the aw valid/ready signals, since this is what should
trigger the initial alignment of the demux.

Split the demux into one variable that counts how many
cycles, and another one that just tracks the phase of the demux.

Trying to do both with one variable is too hard to reason through,
at least for me.

This cleans up a couple more edge cases.