FIXME: flip polarity of reverse for downconverter case
Not sure what is driving this, but, empirically the bytes
are being swapped on the read path.

Unfortunately the stride converter only has the ability to
swap both at the same time, not one or the other.

Could be possible that maybe the solution is to write the data
into the memory in reverse order of what is expected? not sure.