FIXED: the problem was that the write-side mux polarity is flipped
went through and checked the written values in memory, and it
looks like I had them flipped. This should fix that.

More testing is needed, but at this point I think a pretty rich
set of corner cases have been hit from my test bench perspective.

Not sure what the status of this RFC should be -- the Upconverter
doesn't work as-is for AXI, mainly because you're allowed to
initiate writes at unaligned addresses from what the converter
expects, but the converter was oblivious to addresses.

The changes I've made bring the address data into the converter
so it can setup the mux appropriately.

Only tested for 32->64 bit conversions, which is the case of
interest for me, but, I'll update this if I find any more corner
cases.