FIX: fix wishbone arbiter
Commit 07184d37df9921fa2bf729f131d276402bbb3605 breaks designs
in my CI testbench. The intent of the commit was to change the
wishbone arbiter policy for "everyone" to solve a corner case,
but in fact, it breaks other corner cases.

This fix does the following:

- makes the legacy arbiter choice the default, so we don't break
designs that currently work just fine

- expose an `arbiter_policy` flag available on `soc.bus` which
can be set to override the default policy for designs that
require a different arbiter policy to work around a particular
use case.

Even if this PR isn't what one would go with syntatically, I
would strongly suggest either making the change optional or
figuring out why changing the wishbone arbiter policy breaks
existing designs. Really chasing down all the implications of
a bus policy change like this is a big project that could take a lot
of time, especially on SoCs with a single CPU core that is expected
to "own" the bus.

@mtdudek it's a big change to switch the default shared
fabric arbiter policy for an entire design ecosystem!