add a non-registered option for CSR responders
axi-lite would require that the data be valid at the same time
that the `ready` bit is asserted. Putting the data out one cycle
after the register is selected would break this requirement.

This is part of a fix to ensure that `.re`/`.we` signals are
asserted for just one cycle on axi-lite CSRs.