ensure that the address is valid for only one cycle on reads
In order to ensure that `.we` is asserted for only one cycle,
the address may only be valid for one cycle (otherwise `.we` is
asserted for longer than one cycle, since it is trivially decoded
from the address by the CSR mechanism).

This commit changes `START-TRANSACTION` and `SEND-READ-RESPONSE`
to assert `port_adr.eq(0)` in all situations except for when `r.valid`
is asserted.

There are also some changes from PR #1506 polluting this request,
because it depends upon this PR being merged for the simulations I'm
running to work. Presumably this would squash into the master branch
once the PR is accepted.