fix double-read/write on axi2axilite
axi2axilite will initiate two cycles for both reads and writes
when a single cycle is requested by the axi bus. This is harmless
if repeated reads and writes to a CSR don't affect the system state,
however, it is technically incorrect.

This fixes this problem:

1. ax_beat.ready is being suppressed for some reason. Not sure
why, but it causes the logic to think that the beat did not
complete on time, causing the beat initiator to stay on for longer
than it should.

2. ax_lite.ar.valid is being asserted for an extra cycle. The
valid signal is being cut off on `~_cmd_done`. However, _cmd_done
does not change state until the following cycle, so, the `valid`
is asserted longer than it should be. This change takes the condition
for `_cmd_done` and folds it into the current valid setting, allowing
the `valid` signal to drop on time.

The combination of these two changes prevents the axi-lite bus
from double-initiating reads and writes.

I do have concerns that this could break in the case that a burst is
initiated from the host. However, my test model CPU won't do burst
writes or reads to the CSR bus because the region is marked as
uncached and writes are strictly write-through so they only happen
one cycle at a time.

In general, though, I think this the expected behavior of a CSR.

I'll try to build another model where a RAM is hanging off of the
axi-lite bus in a cached region to see if I can't create some real-world
bursting cycles to test this further, but, maybe someone has a comprehensive
test case for AXI burst to axi-lite beat ready to go?