fix cycle delay + make range logic generic across configs
The mux was actually delayed by one cycle, but this didn't show
up in my earlier testbench because I was simply counting up while
writing data. I now have a "random" (actually LFSR) write, sequential
read test bench that is passing and it picked up this bug. The fix
was to hoist the demux computation into the case statement, so it
slightly increases that critical path but I think it's probably OK.

I also formatted the demux_val priming logic so it is no longer
specific to my system config, it's now generic so this could be
considered a PR for mainlining.