clean up combinational loops in axi_lite_to_simple converter
The use of a statement such as:

If(a,
   b.eq(1)
)

generates verilog that looks something like this:

always @(*) begin
   b <= 0;
   if (a) begin
      b <= 1;
   end
end

the problem is the `b <= 0;` statement at the top of the
always @(*) begin block. If later on b turns out to impact a,
it will trigger the sensitivity list again, and you end up
with an infinite loop in simulation.

This commit hoists `b` out of the `always @` block by sticking
it into a `comb` prefix, avoiding the initialization loop.