liblitedram/sdram_spd: use temp_len when reading SPD
This change should change anything from the point of view of
`sdram_read_spd` caller, but it makes it so less I2C reads are actually
hapenning.

With `len` we read too many bytes and write them to the `buf`.
In subsequent iterations we overwrite those bytes as all counters are
being updated by the `temp_len`. Nothing terrible, but too many
bytes were being read.

Signed-off-by: Michal Sieron <msieron@antmicro.com>