csr: add ability to express clusters
This adds the ability to express clusters of registers. This can be very
useful for peripherals with repeated groups of registers.

A `CSRCluster` class, emulating a simplified array, is added.
Peripherals can add items to the cluster, and the CSRs will, for JSON,
CSV, and HTML, be collected as normal.

For SVD, the code is extended to be able to handle outputting clusters
directly in the XML structure. This way, code generation tools like
`svd2rust` can generate nested array structures, making ergonomics
significantly better when working with peripherals.

The code supports nesting of `CSRClusters` too.