tools/litex_json2dts_zephyr: add mapped SPI flash for XIP support
This allows to execute a firmware loaded in the SPI flash.

No support for partitions yet: this does not yet support cohabitating the
bitstream and firmware yet, so the bitstream has to be loaded to CRAM
directly for now (default for --load).

This allows loading larger Zephyr examples, and was required to fit
the shell on the Crosslink-NX board.

See https://github.com/litex-hub/zephyr-on-litex-vexriscv/pull/13