litex_setup.py: Disable auto-update of litex_setup for CI
This adds a new config option to litex_setup (--disable-auto-update) to disable
the auto-update mechanism. It further changes the CI flow to use that flag.

Why? The auto-update is a problem for the CI workflow because it will override
any (especially branch-specific) changes to the litex_setup script. Then CI
is testing the wrong litex_setup script.