cpu/rocket: change load path of generated verilog files
Since rocket-chip upstream switched to mill build, it can not generate
a single verilog file, instead, it generate a lots of file in build
directory. To save space, we do deduplication for the common file.
This means that we need to load verilog file from different directory
(common code and config specific).

Add all the path needed for generated verilog files.

Signed-off-by: Inochi Amaoto <inochiama@outlook.com>