bios: add BOOTP support to libliteeth
This adds BOOTP support as per RFC 951 and RFC 1532. Since most recent
BOOTP servers seem to require the magic cookie defined in RFC 1533, this
is also included.

N.b.: the secs field is set to the uptime, even though RFC 1532
specifies this as the time since the first BOOTREQUEST. Since the BIOS
usually won't be running for that long and the field is often ignored by
servers, this felt like a reasonable compromise.

Signed-off-by: Matthias Breithaupt <m.breithaupt@vogl-electronic.com>