soc/cores/clock/intel_agilex: refactor the Altera Agilex PLL core
On Agilex families, the "derive_pll_clocks" SDC command is deprecated,
hence we need to provide proper "create_generated_clock" statements
instead.

To achieve this, we need to provide the platform object when
creating the PLL object so we can reference the toolchain string
needed to add the additional SDC commands.

Also, because we have the device string from the platform object we can
derive the speedgrade from there (currently only done on Agilex 3).

Signed-off-by: Steffen Persvold <spersvold@gmail.com>